Isolate audit refresh cache gate
Keep the full-pipeline no-reopen test in the serialized Console collection so its required cache initialization cannot race cache command tests.
